+++ cmd/vcfWebAnnotate/vcfWebAnnotate.go
@@ -24,62 +24,35 @@ func usage() {
 	flag.PrintDefaults()
 }
 
-// Example of JSON layout can be found in the link below
-// http://bioinfo.hpc.cam.ac.uk/cellbase/webservices/rest/v4/hsapiens/genomic/variant/chr1%3A878884%3AC%3AT,chr1%3A878917%3AT%3AA,chr1%3A878920%3AT%3AA,chr1%3A878991%3AGTGTT%3AG,chr1%3A879229%3AA%3AT,chr1%3A879231%3AA%3AC,chr1%3A879897%3AT%3AC,chr1%3A879957%3AG%3AT/annotation?assembly=grch38
+func vcfWebAnnotate(data <-chan vcf.Vcf, header vcf.Header, outfile io.Writer, batchSize int, numBuffers int) {
+	header = addAnnotationHeader(header)
+	vcf.NewWriteHeader(outfile, header)
+	filledBufChan := make(chan []vcf.Vcf, numBuffers)
+	emptyBufChan := make(chan []vcf.Vcf, numBuffers)
 
-type Responses struct { // the response for each variant queried
-	Responses []Response `json:"response"`
-}
+	for i := 0; i < numBuffers-1; i++ {
+		emptyBufChan <- make([]vcf.Vcf, 0, batchSize) // send all but 1 buffer to empty
+	}
+	buf := make([]vcf.Vcf, 0, batchSize)
 
-type Response struct { // all the results for a single variant (e.g. multiple transcripts)
-	Results []Result `json:"result"`
-}
+	go queryWorker(filledBufChan, emptyBufChan, outfile)
 
-type Result struct { // data on a particular variant on a particular transcript
-	Chr             string          `json:"chromosome"`
-	Start           int             `json:"start"`
-	Ref             string          `json:"reference"`
-	Alt             string          `json:"alternate"`
-	SnpId           string          `json:"id"`
-	ConsequenceType string          `json:"displayConsequenceType"`
-	Consequences    []Consequence   `json:"consequenceTypes"`
-	PopAlleleFreqs  []PopAlleleFreq `json:"populationFrequencies"`
-	// Fields below exist but are not collected for this cmd
-	// conservation
-	// geneExpression
-	// geneTraitAssociation
-	// geneDrugInteraction
-	// cytoband
-	// repeat
-}
+	for v := range data { // read vcfs until you have a full batch then send for annotation
+		if len(buf) == batchSize {
+			filledBufChan <- buf
+			buf = <-emptyBufChan
+			buf = buf[:0]
+		}
+		buf = append(buf, v)
+	}
 
-type PopAlleleFreq struct { // pop allele frequency from gnomAD and 1kgp
-	Study      string  `json:"study"`
-	Population string  `json:"population"`
-	RefAf      float64 `json:"refAlleleFreq"`
-	AltAf      float64 `json:"altAlleleFreq"`
-}
+	if len(buf) > 0 { // send any variants remaining in the buffer
+		filledBufChan <- buf
+	}
+	close(filledBufChan)
 
-type Consequence struct { // information about variant effect
-	GeneName          string            `json:"geneName"`
-	GeneId            string            `json:"ensemblGeneId"`
-	TranscriptId      string            `json:"ensemblTranscriptId"`
-	Strand            string            `json:"strand"`
-	Biotype           string            `json:"biotype"`
-	ProteinAnnotation ProteinAnnotation `json:"proteinVariantAnnotation"`
-}
-
-type ProteinAnnotation struct { // variants effect on protein
-	Pos                int                  `json:"position"`
-	Ref                string               `json:"reference"`
-	Alt                string               `json:"alternate"`
-	SubstitutionScores []SubstitutionScores `json:"substitutionScores"`
-}
-
-type SubstitutionScores struct { // sift & polyphen scores
-	Source      string  `json:"source"`
-	Score       float64 `json:"score"`
-	Description string  `json:"description"`
+	for _ = range emptyBufChan { // stall until queryWorker is finished
+	}
 }
 
 func queryWorker(filledBufChan <-chan []vcf.Vcf, emptyBufChan chan<- []vcf.Vcf, outfile io.Writer) {
@@ -196,37 +169,6 @@ func addAnnotationHeader(header vcf.Header) vcf.Header {
 
 	header.Text = append(header.Text, savedHeader...)
 	return header
-}
-
-func vcfWebAnnotate(data <-chan vcf.Vcf, header vcf.Header, outfile io.Writer, batchSize int, numBuffers int) {
-	header = addAnnotationHeader(header)
-	vcf.NewWriteHeader(outfile, header)
-	filledBufChan := make(chan []vcf.Vcf, numBuffers)
-	emptyBufChan := make(chan []vcf.Vcf, numBuffers)
-
-	for i := 0; i < numBuffers-1; i++ {
-		emptyBufChan <- make([]vcf.Vcf, 0, batchSize) // send all but 1 buffer to empty
-	}
-	buf := make([]vcf.Vcf, 0, batchSize)
-
-	go queryWorker(filledBufChan, emptyBufChan, outfile)
-
-	for v := range data { // read vcfs until you have a full batch then send for annotation
-		if len(buf) == batchSize {
-			filledBufChan <- buf
-			buf = <-emptyBufChan
-			buf = buf[:0]
-		}
-		buf = append(buf, v)
-	}
-
-	if len(buf) > 0 { // send any variants remaining in the buffer
-		filledBufChan <- buf
-	}
-	close(filledBufChan)
-
-	for _ = range emptyBufChan { // stall until queryWorker is finished
-	}
 }
 
 func main() {
